Fixes various issues Docker Events no longer used Account for removal of subscription for user with no customer record

+++ lib/cashier.rb
@@ -23,8 +23,10 @@ module Pushbit
     end
 
     def unsubscribe
-      remove_repo_from_subscription
-      subscription.destroy
+      if user.has_customer?
+        remove_repo_from_subscription
+        subscription.destroy
+      end
     end
 
     def subscription
